# Conflicts:
#	python/sim/dfb.py

+++ examples/matmul-tutorial/step_1_single_node_single_tile_block.py
@@ -188,7 +188,7 @@ try:
     b = torch.randn((K, N), dtype=torch.bfloat16)
     c = torch.randn((M, N), dtype=torch.bfloat16)
 
-    expected_y = torch.relu(a @ b + c)
+    expected_y = torch.relu(a.float() @ b.float() + c.float()).to(torch.bfloat16)
 
     a = from_torch(a)
     b = from_torch(b)
